Updated schedule
Changed schedule to 20 minutes past the hour, every 3 hours, between 08:00 and 22:59, Monday through Friday. This is because the first schedule of the morning sometimes fails to run. By making the first run earlier, the second run is brought forward. Schedule now continues to 11PM to provide for those based in NYC.